+++ opentelemetry-sdk/src/trace/mod.rs
@@ -80,8 +80,6 @@ mod tests {
             span.add_event("test-event".to_string(), vec![]);
         });
 
-        provider.force_flush();
-
         // Assert
         let exported_spans = exporter
             .get_finished_spans()
@@ -112,8 +110,10 @@ mod tests {
         span.set_attribute(KeyValue::new("attribute1", "value1"));
         span.add_event("test-event".to_string(), vec![]);
         span.set_status(Status::error("cancelled"));
-        drop(span);
-        provider.force_flush();
+        span.end();
+
+        // After span end, further operations should not have any effect
+        span.update_name("span_name_updated");
 
         // Assert
         let exported_spans = exporter
@@ -150,7 +150,6 @@ mod tests {
         span.add_event("test-event".to_string(), vec![]);
         span.set_status(Status::Ok);
         drop(span);
-        provider.force_flush();
 
         // Assert
         let exported_spans = exporter
@@ -182,22 +181,18 @@ mod tests {
 
         let mut links = Vec::new();
         for _i in 0..(DEFAULT_MAX_LINKS_PER_SPAN * 2) {
-            links.push(Link::new(
-                SpanContext::new(
-                    TraceId::from_u128(12),
-                    SpanId::from_u64(12),
-                    TraceFlags::default(),
-                    false,
-                    Default::default(),
-                ),
-                Vec::new(),
-            ))
+            links.push(Link::with_context(SpanContext::new(
+                TraceId::from_u128(12),
+                SpanId::from_u64(12),
+                TraceFlags::default(),
+                false,
+                Default::default(),
+            )))
         }
 
         let span_builder = SpanBuilder::from_name("span_name").with_links(links);
         let mut span = tracer.build(span_builder);
         span.end();
-        provider.force_flush();
 
         // Assert
         let exported_spans = exporter
@@ -233,7 +228,6 @@ mod tests {
         span.add_event("test event again, after span builder", Vec::new());
         span.add_event("test event once again, after span builder", Vec::new());
         span.end();
-        provider.force_flush();
 
         // Assert
         let exported_spans = exporter